pwm_lut: Output driver linearisation LUT calculator
For use with CMOS-driven rc filtered PWM outputs.

Computes the table of DAC output voltage vs duty cycle and extant
voltage on the DAC output. Vo is a function of Rds(n) and Rds(p) of
the totem pole transistors, which themselves are a function of Ids and
intrinsic parameters.

Todo: add commandline handling